refac:
- rename `VaemDriver.py` to `driver.py`.
- rename `dataTypes.py` to `config.py`
- rename `vaemHelper.py` to `utils.py`
- update imports in __init__.py to reflect updated structure.
- make `set_valve_setting` and `read_valve_setting` private methods and rename `set_all_valve_settings` and `read_all_valve_settings` to `set_valve_settings` and `read_valve_settings` respectively. Users should now only use these.
