+++ src/pyvaem/__init__.py
@@ -1,2 +1,2 @@
-from .VaemDriver import VaemDriver
-from .dataTypes import VaemConfig
+from pyvaem.config import VaemConfig, ValveSettings
+from pyvaem.driver import VaemDriver
