--- src/pyvaem/dataTypes.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class VaemConfig:
-    ip: str
-    port: int
-    slave_id: int
