+++ src/pyvaem/driver.py
@@ -3,15 +3,14 @@ from typing import Callable, Concatenate, ParamSpec, TypeVar
 
 from pymodbus.client import ModbusTcpClient as TcpClient
 
-from pyvaem.dataTypes import VaemConfig
-from pyvaem.vaemHelper import (
+from pyvaem.config import VaemConfig, ValveSettings
+from pyvaem.utils import (
     VaemAccess,
     VaemControlWords,
     VaemDataType,
     VaemIndex,
     VaemOperatingMode,
     VaemRegisters,
-    ValveSettings,
     create_controlword_registers,
     create_select_valve_registers,
     create_setting_registers,
@@ -215,7 +214,7 @@ class VaemDriver:
 
     ### VALVE SETTINGS OPERATIONS ###
     @clear_and_raise_error
-    def set_valve_setting(
+    def _set_valve_setting(
         self, valve_id: int, setting: VaemIndex, value: int
     ) -> VaemRegisters:
         valid_range = vaem_ranges.get(setting.name)
@@ -235,7 +234,7 @@ class VaemDriver:
         )
         return self._transfer_vaem_registers(data)
 
-    def set_multiple_valve_settings(
+    def set_valve_settings(
         self, valve_id, settings: ValveSettings | dict[str, int] | None
     ):
         """Configure settings for a specific valve. This method allows setting various
@@ -260,7 +259,7 @@ class VaemDriver:
             return
 
         for setting, value in valve_settings.to_enum_dict().items():
-            self.set_valve_setting(valve_id, setting, value)
+            self._set_valve_setting(valve_id, setting, value)
 
     @clear_and_raise_error
     def save_settings(self) -> VaemRegisters:
@@ -276,7 +275,7 @@ class VaemDriver:
         return self._transfer_vaem_registers(data)
 
     @clear_and_raise_error
-    def read_valve_setting(self, valve_id, setting: VaemIndex) -> VaemRegisters:
+    def _read_valve_setting(self, valve_id, setting: VaemIndex) -> VaemRegisters:
         """Read settings for a specific valve."""
         # Check if parameter is actually a setting
         if setting.name not in vaem_parameters:
@@ -289,14 +288,14 @@ class VaemDriver:
         )
         return self._transfer_vaem_registers(data)
 
-    def read_all_valve_settings(self, valve_id: int):
+    def read_valve_settings(self, valve_id: int):
         """Read all settings for a specific valve and returns them as a ValveSettings object"""
         if valve_id not in range(1, 9):
             raise ValueError("Valve_id must be between 1-8")
 
         settings: dict[str, int] = {}
         for setting in vaem_parameters:
-            value = self.read_valve_setting(valve_id, getattr(VaemIndex, setting))
+            value = self._read_valve_setting(valve_id, getattr(VaemIndex, setting))
             if value is not None:
                 settings.update({setting: value.transferValue})
